PTERO-T-1 Added tests for reveal method
